+++ apps/backoffice-v2/src/domains/entities/hooks/mutations/useRejectTaskByIdMutation/useRejectTaskByIdMutation.tsx
@@ -6,7 +6,7 @@ import { workflowsQueryKeys } from '../../../../workflows/query-keys';
 import { Action } from '../../../../../common/enums';
 import { useFilterId } from '../../../../../common/hooks/useFilterId/useFilterId';
 
-export const useRejectTaskByIdMutation = (workflowId: string, postUpdateEventName?: string) => {
+export const useRejectTaskByIdMutation = (workflowId: string) => {
   const queryClient = useQueryClient();
   const filterId = useFilterId();
   const workflowById = workflowsQueryKeys.byId({ workflowId, filterId });
@@ -19,7 +19,6 @@ export const useRejectTaskByIdMutation = (workflowId: string, postUpdateEventNam
         body: {
           decision: Action.REJECT,
           reason,
-          postUpdateEventName,
         },
       }),
     onMutate: async ({ documentId, reason }) => {