+++ apps/backoffice-v2/src/main.tsx
@@ -1,6 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
+import '@total-typescript/ts-reset';
+
 import '@ballerine/ui/dist/style.css';
 import '@fontsource/inter';
 
@@ -14,7 +16,7 @@ export const TOAST_DURATION_IN_MS = 1000 * 3;
 
 const rootElement = document.getElementById('root');
 
-// Avoid race conditions when using the mock server.
+/// Avoid race conditions when using the mock server.
 const prepare = async () => {
   if (env.VITE_MOCK_SERVER) {
     const { worker } = await import('./lib/mock-service-worker/browser');