+++ services/workflows-service/scripts/seed.ts
@@ -18,7 +18,7 @@ import { generateKycSessionDefinition } from './workflows/kyc-email-process-exam
 import { env } from '../src/env';
 import { generateKybKycWorkflowDefinition } from './workflows/kyb-kyc-workflow-definition';
 import { generateBaseTaskLevelStates } from './workflows/generate-base-task-level-states';
-import { generateBaseCaseLevelStates } from './workflows/generate-base-case-level-states';
+import { generateBaseCaseLevelStatesAutoTransitionOnRevision } from './workflows/generate-base-case-level-states';
 import type { InputJsonValue } from '../src/types';
 import { generateWebsiteMonitoringExample } from './workflows/website-monitoring-workflow';
 import { generateCollectionKybWorkflow } from './workflows/generate-collection-kyb-workflow';
@@ -549,7 +549,7 @@ async function seed(bcryptSalt: string | number) {
   // KYB Manual Review (workflowLevelResolution true)
   await client.workflowDefinition.create({
     data: {
-      ...baseReviewDefinition(generateBaseCaseLevelStates()),
+      ...baseReviewDefinition(generateBaseCaseLevelStatesAutoTransitionOnRevision()),
       id: kybManualMachineId,
       config: {
         workflowLevelResolution: true,
@@ -871,7 +871,7 @@ async function seed(bcryptSalt: string | number) {
         context: {
           documents: [],
         },
-        states: generateBaseCaseLevelStates(),
+        states: generateBaseCaseLevelStatesAutoTransitionOnRevision(),
       },
     },
   });
@@ -968,16 +968,6 @@ async function seed(bcryptSalt: string | number) {
       where: {
         workflowDefinitionId: { in: ['kyb_with_associated_companies_example'] },
         businessId: { not: null },
-        state: {
-          in: [
-            'manual_review',
-            'approved',
-            'revision',
-            'rejected',
-            'pending_kyc_response',
-            'pending_kyb_response',
-          ],
-        },
       },
     },
     project1.id,