+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,47 @@
 # kyb-app
 
+## 0.1.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-browser-sdk@0.5.45
+
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.44
+  - @ballerine/workflow-browser-sdk@0.5.44
+
+## 0.1.44
+
+### Patch Changes
+
+- Ver bump
+- Updated dependencies
+  - @ballerine/blocks@0.1.27
+  - @ballerine/common@0.7.43
+  - @ballerine/ui@0.3.29
+  - @ballerine/workflow-browser-sdk@0.5.43
+
+## 0.1.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.42
+  - @ballerine/workflow-browser-sdk@0.5.42
+
+## 0.1.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/blocks@0.1.26
+
 ## 0.1.41
 
 ### Patch Changes