+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ballerine/eslint-config
 
+## 1.0.7
+
+### Patch Changes
+
+- added rules to enforce spacing
+
+## 1.0.6
+
+### Patch Changes
+
+- Ver bump
+
 ## 1.0.5
 
 ### Patch Changes